Modify add and remove functionality
